requirements.txt
streamlit
google-generativeai
pandas
plotly
python-dotenv
openpyxl
